Move model types from balances to defi
[ui tests]
